Updated the file structure such that code is more readable
2 new folders added `subsystemHeaders` and `subsystemFiles`

`subsystemHeaders` contains global declarations of devices in `globals.hpp` and will contain global declarations of functions in all files that are named after the robots subsystems.

`subsystemFiles` contains initialisations of device ports (other parameters are initialised in `main.cpp` but that is subject to change). Function parameters are initialised in the `.cpp` files named after the robots subsystems. These functions will be used in `main.cpp` to execute that actual code so that code is less cluttered and organised into files with names that make sense.