Added Pneumatics and Drivetrain fixes
Declared all pneumatics (not sure if the Mogo Mech has 1 or 2 pistons so I just added 2), made a function to toggle the wall stake claw though a more sophisticated macro is probably due. Also fixed the drivetrain motor names and weather or not they're reversed to be inline with the triangular setup of the motors.